sorted redirects, removed unnecessary component for page
